code refactor
- fix typos
- reset ws reconnect count on connection open
- run files inspection of files upload
- remove VideoStateProgress as wrapping structure for VideoStateProgress

+++ modules/backend/src/features/video/inspect/dto.rs
@@ -8,7 +8,7 @@ use crate::features::video::model::{AudioStream, MediaMetadata, VideoStream};
 pub struct VideoInspectionResponse {
   pub original_file_name: String,
   pub file_size_bytes: i64,
-  pub duration_seconds: f32,
+  pub duration_seconds: f64,
   pub video_streams: Vec<VideoStream>,
   pub audio_streams: Vec<AudioStream>,
 }
