+++ modules/backend/src/features/video/model.rs
@@ -1,15 +1,15 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize)]
+#[derive(Clone, Debug, Serialize)]
 pub struct MediaMetadata {
   pub path: String,
-  pub duration_seconds: f32,
+  pub duration_seconds: f64,
   pub file_size_bytes: i64,
   pub video_streams: Vec<VideoStream>,
   pub audio_streams: Vec<AudioStream>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct VideoStream {
   pub id: String,
   pub bit_rate: i32,
@@ -19,7 +19,7 @@ pub struct VideoStream {
   pub width: i32,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct AudioStream {
   pub id: String,
   pub bit_rate: i32,
